requirements.txt
-i https://pypi.org/simple
alembic==1.10.2
aniso8601==9.0.1
click==8.1.3
faker==18.3.0
flask==2.2.3
flask-cors==3.0.10
flask-migrate==4.0.4
flask-restful==0.3.9
flask-sqlalchemy==3.0.3
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
honcho==1.1.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.4
markupsafe==2.1.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
setuptools==67.6.0
six==1.16.0
sqlalchemy==2.0.7
sqlalchemy-serializer==1.4.1
typing-extensions==4.5.0 ; python_version >= '3.7'
werkzeug==2.2.3
zipp==3.15.0 ; python_version >= '3.7'
